pik_clock: make divider value check limit as configurable parameter
PIK clock divider value limit checking is hardcoded to 16 with
4-bit assumption. However many mobile & infra platforms has
5-bit divider.

This patch fixes the issue by adding a configurable module parameter
so that the platform's config file can set the divider bitfield size
based on platform.

Additionally, if platform does not provide module configuration data,
then the divider limit is set to a default value of 5-bits.

Change-Id: I70103ec84eacc2d26be999ca6524878bff51dd9d
Signed-off-by: Manoj Kumar <manoj.kumar3@arm.com>