sensor: Add initial support for asynchronous drivers
With this patch the sensor HAL can check a pending
status returned by non-blocking drivers which defer
the return of the requested value until a later stage.
However, the handling of such situations is not yet
supported.
In addition, the allocated memory for the elements
at init stage is zeroed, in preparation for fully
supporting the above feature.

Change-Id: Ib58a96ed74e31155e9b0539d25c8a6d6afdb7d60
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>