system_power: Fix extended PPU support
This patch fixes an issue in the support for extended PPU handling in
system_power, where the ext_ppu_apis field represented a pointer to an
API structure, whilst being initialised like a pointer to an array of
pointers to API structures.

This manifested itself in a crash whenever the module tried to iterate
over the array to perform power management tasks.

Change-Id: I81d071d82f8ed0f677f585a83670ed519cdb8d26
Signed-off-by: Chris Kay <chris.kay@arm.com>