ppu_v0: Ensure bound_id is set when bound by system_power
For PPUs being bound to by system_power instead of power_domain, the
bound_id field was not being updated to reflect the binder. This caused
issues later down the line when trying to use the power domain driver
API of the domain, because the API binding would be skipped if bound_id
was not assigned.

Change-Id: Ife239724026d5d8cf41626811953a4eefb12bf10
Signed-off-by: Chris Kay <chris.kay@arm.com>