module: Add debug module
This patch adds a debug module among the core modules. The debug module
provides an abstraction layer between SCMI Power Domain and the Power
Domain HAL. It allows for a driver to be plugged in and handle platform
specific operations to be performed.

The debug module also makes it possible to handle requests to
enable/disable debug infrastructure from multiple sources, such as from
OS and hardware.

Change-Id: Ie2d4482be30b4baae329e55dfc266f3877ad0463
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Signed-off-by: Raphael Gault <raphael.gault@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>