juno/module: Add debug driver
This patch adds the driver for Juno to support debug infrastracture
on the platform.

This driver is responsible for handling debug requests coming
from the hardware, i.e. when a debugger is attached and
a connection is requested.
Also, this driver implements the mod_debug_driver API
to enable/disable DBGSYS and additional areas of the SoC
required to properly enable the debug infrastructure.

Change-Id: I363edc84cd23a12153891f4549fe10d8e349ff9a
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Signed-off-by: Raphael Gault <raphael.gault@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>