n1sdp: fix n1sdp_sensor hang by moving alarm ID to config file
n1sdp_sensor module uses timer alarm API for periodically measuring the
sensor values. So the module binds to timer module in the bind stage
using FWK_ID_SUB_ELEMENT call. So when a library is built, the value
calculated by the macro will be taken based on that build environment.

However, in later stage if modules added/removed in firmware.mk, the
position of the timer module will be changed and since a library module
will not be built in normal build phase, it is unaware of the change in
position and hence passes wrong element id to timer during bind. This
makes the code to fail.

As a workaround, this patch moves the id calculation to config file as a
module configuration data so that the value is generated in every build.
The library will read the data passed by the config file and not uses
fixed value.

Change-Id: Id7acf2e6899d7e0f0c106994fd0303d67a596071
Signed-off-by: Manoj Kumar <manoj.kumar3@arm.com>