juno_xrp7724: add support for shutdown and cold reset
This patch adds an interface to perform a shutdown and a cold
reset of the Juno board.
These features are managed by the XRP7724 PMIC on the Juno board,
and are performed by writing two of its GPIOs via an I2C communication
bus.

On the Juno board, one of the GPIOs is programmed to select if the PMIC
will perform a shutdown or a cold reset operation. Setting the second
GPIO will trigger the operation selected with the first GPIO.

Change-Id: Idecb92034bd4f26024ba631565a919fab44b7c8d
Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>