n1sdp: return link training status based on CCIX root port
This patch return success status for CCIX root port even if link
training fails as SCP should continue even if the slot is not
connected with any cards. However it should return error code
for link training failure on standard PCIe root port as it is
connected to on-board PCIe switch.

Change-Id: I8a8cc92293efb5b1e1038a1c7cf3a02293eb0546
Signed-off-by: Manoj Kumar <manoj.kumar3@arm.com>