fwk/notification: Allow use of provided source_id for notify
Currently, when a notification is sent, the source identifier
of such notification is set up by the framework to the entity
of the target for the event being processed.

With this change, the framework does not overwrite anymore
the source identifier of the notification event if the provided
source identifier is a valid entity identifier.

This change allows the use case where a module processing the
current event calls another module API and the called entity
needs to send a notification. In this scenario the called module
can simply provide a valid source_id for the outgoing notification
event.

Change-Id: I9d696b46ad656374387e97cfce6dba95a41fd182
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>