system_power: Move to dynamic last core identifier
To ensure proper operation on system resume, the last
core on system suspend being turned off needs to be the
first one to be turned on. This requires the identifier
of the power domain of that core to be fetched and saved
before system suspend.

This patch replaces the fixed pre-defined power domain
for the primary cpu with the one provided by the power
domain module via get_last_core_pd_id.

Change-Id: I7538bcb1dd95d99faf65a425c58c6c3a2e968a00
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>