I2C: Add FWK_PENDING to the I2C HAL interface
This patch adds support for returning FWK_PENDING from an interface
with the I2C HAL module. The function returns FWK_PENDING when the
transaction has been submitted but not completed. The completion will
be notified later by a response event.

At the same time, drivers and clients are updated to take into account
these changes.

Change-Id: I87d83e005ebf1b76f1707519131ec3d6eb5344b7
Co-authored-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>
Signed-off-by: Jim Quigley <jim.quigley@arm.com>
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>