juno_xrp7724: Add temperature sensor support
This patch adds the support for the temperature sensor of the
XRP7724 component of the Juno board.

It provides an implementation for the sensor driver API and
uses the I2C HAL API to make sensor requests to the XRP7724
component. It adds the temperature sensor to the configuration
of the XRP7724 module on Juno.

Change-Id: Ic6d1e47328b118eb648ba2c4869cde9f1b3866fc
Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>