pl011: Add powerup/powerdown handlers
This patch introduces a configurable power domain identifier
such that the driver can be informed of power state changes.
This is achieved through notifications for pre_state and
post_state power changes.
With this power domain identifier the module can receive notifications
and action accordingly. The specific action is not yet implemented.

Change-Id: I7751b6de5899a2c65aec3aef3e5e4b8c1d0cd07c
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>