arch/armv7-m: Split exception table from handlers
This commit splits out the exception table from the handlers it links
to. We need to do this in order to wrap the reset handler function,
which is needed for Juno (see the next commit).

The current situation is that, when compiling the exception table, the
compiler can determine the address of the reset function statically.
This means that by the the time the linker has had a chance to wrap the
reset handler function, the link between the exception table and the
function has already occurred, making it impossible for the linker to
link to our wrapper function instead.

We need to alter this situation so that when the linker comes to a
reference to the reset handler, it can link it to the wrapped handler
instead. To do this, we need to make sure that the compiler cannot
determine the location of the function by itself, which requires
putting the table and the handler in separate compilation units.

Change-Id: Id17ddea28a3f26323bef45f4ce8961d50b866c59
Signed-off-by: Chris Kay <chris.kay@arm.com>