juno/scp_romfw_bypass: Ensure system PLLs are locked within MCC timeout
The current approach to resetting the PLLs, which is to use a pre-main
constructor function, is not infallible - if we statically allocate
enough data, eventually the time it takes for the C runtime to
initialize exceeds the amount of time that the motherboard
microcontroller gives us to ensure the system PLLs are locked.

The workaround for this is to instead wrap the core's reset handler.
This way we ensure that the PLLs have been reset before we do any time-
consuming work.

Change-Id: Ie74df40c8de2ea353ec048c447c85cc7d013b94b
Signed-off-by: Chris Kay <chris.kay@arm.com>