fwk_thread: Clarify support for delayed responses
Currently, delayed responses can be retrieved only
when support for multi-threading is enabled.

The documentation for fwk_thread mistakenly claims
that there is an implementation for pulling delayed
responses.

This patch makes explicit that delayed responses
are not supported in single-threading by updating the
relevant documentation and implementing a clear
always-return FWK_E_SUPPORT.

Change-Id: I0cf340eb537384af27bde5f06b50d844af6180b3
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>