fwk: Add ring buffer data structure
This commit adds a ring buffer data structure type, also known as a
circular buffer.

Ring buffers are a useful tool for making the most use out of a
fixed-size buffer, where the practical beginning and end of the
buffer may exist at arbitrary points within it.

Reads and writes are treated as though each end is connected,
allowing data to read or write "off the end", and is typically
used to serve as a first-in, first-out buffer.

Change-Id: I68fa8d96cf9bb62797c3ce3a661176673c017b7a
Co-authored-by: Brett Warren <brett.warren@arm.com>
Signed-off-by: Brett Warren <brett.warren@arm.com>
Signed-off-by: Chris Kay <chris.kay@arm.com>