fwk/test: Refactor framework test build system
This refactoring cleans up the framework test makefiles, reducing the
amount of boilerplate needed to create framework tests.

The build system logic for the test framework has been adjusted so that
framework source files are recompiled for each test case. This ensures
that build flags can be set on a per-test-case basis, where currently
the framework is built only once with only one set of build flags. This
means test cases can selectively enable and disable the necessary
features.

In addition, the framework is compiled in its entirety for each test
suite. This means that test suite no longer need to selectively identify
the source files required to successfully build.

Change-Id: I0d4c10b5903d47ee5f81bd70dc453331dd4d236f
Signed-off-by: Chris Kay <chris.kay@arm.com>