psu: Add support for asynchronous drivers
This commit adds the extra handling required to support drivers that may
pend certain requests, as opposed to acting on them immediately.

This extra support is intended for drivers that might otherwise busy-
wait for extended periods, such as when they do not immediately know the
result of an operation they have begun.

Change-Id: I39aa9e4129d833044bd0e87b324c701d59a6908f
Signed-off-by: Chris Kay <chris.kay@arm.com>