cmn600: expose API to read chip information from platforms
CMN-600 supports multichip operation and for configuring the
routing tables it requires the chip ID information from the
platform as routing differs based on chip ID.

This patch exposes a common API type which can be implemented
in platform modules to retrieve the information.
The module ID and API ID where the API is implemented is passed
as config parameter to CMN-600.

For N1SDP platform it implements the API in n1sdp_system module
for getting the chip ID. For SGI575 & RDN1E1 platforms it
leaves the module ID and API ID as FWK_ID_NONE in config_cmn600.c
in respective platforms so that the API can be implemented in
platform specific way.

Change-Id: I96867c6cd18621fbc9ea4eccddcce0ace535d2c2
Signed-off-by: Manoj Kumar <manoj.kumar3@arm.com>