module: add system id information module
Add System ID Information hardware abstraction module which provides an
generic interface for modules that generate the System ID information
and modules that consume this information.

This module expects one register interface driver module (like SID or
SSC) to provide a generic information about the system. A module
requesting for system information can bind to this module to get that
information.

Change-Id: Ifd2ada0b8f8278043dd89d5ef0be0ca7f1f0edad
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>