n1sdp: fix power domain handling in C2C module
This patch cleans the power domain API in C2C module and respective
handling in master and slave chips.

It is not possible for slave chip to report power state transition
to master hence on slave chip, sync calls are used to power ON & OFF
the power domains. For this purpose, C2C is treated as separate
thread on slave chips.

Change-Id: Ie356d19d5d4eeba3a62e53e9907f58c944d391b4
Signed-off-by: Manoj Kumar <manoj.kumar3@arm.com>