mock_sensor: Add module
This patch adds a mocker for sensors.
This module implements the sensor driver API and
defers the get_value requests. This module requires
the timer facilities in order to setup an alarm.
This would emulate a sensor that provides a value
through an ISR and a call to the driver_response API
of the sensor module.

This module can be used to test sensor call flow
in platforms that do not provide asynchronous sensor
drivers.

Change-Id: I37baa428c12d6aa1c3de2f6f534ab5ac158f2352
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>