power_domain: Introduce PENDING within shutdown sequence
Power domain drivers may now return FWK_PENDING is they cannot
process requests synchronously.
This patch introduces a basic handling of such situation during
system shutdown where the sequence is in effect interrupted.

At this stage power domain is not providing any driver response
api (see deferred_response_architecture) so it is expected that only
the last driver is realistically qualified to be asyncronous.
In this scenario, that driver is the one in charge of performing the
last operation to turn the system down.

Change-Id: I74012f7289f932675c2eff184337369fcdc7babd
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>