cmn600: improve cxla discovery logic
There are two improvements that can be made to the existing CXLA node
discovery procedure. First, the CXLA node id need not be passed through
the config data. This was initially necessitated to avoid accessing the
external CXLA node when it is not powered up. This can be improved by
using the device type information from the cross point's (XP) port
connect info register and skip over it during the discovery process and
avoid having to provide the CXLA node id from the config data.

Second, with the improved discovery logic, the external RN-SAM count is
incremented only for the external child nodes which are discovered as
RN-SAM device. With this, RN-F and host RA count need not be decreased
by one.

Change-Id: Ia870ac69045957215c7c2fdbce774a230e6fda00
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>