cmn600: rename unique_ha_ldid_value to unique_remote_rnf_ldid_value
The cmn600 context member variable unique_ha_ldid_value is used to
keep track of the unique LDID value of the remote caching agent (RN-F).
Rename it to unique_remote_rnf_ldid_value to align with the purpose of
the variable. This patch also removes the variable from the cmn600
context structure since it is been used only within the ccix_setup
function.

Change-Id: Ib44cad5671e9717469148cf1d9213ca3118cde1d
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>