cmn600: add rnf count to cmn600 context
Currently the RN-F count is considered equal to the total external
RN-SAM count. This is done with an assumption that all the RN-F nodes in
the mesh has no built-in SAM (RN-F ESAM type). If a mesh contains RN-F
nodes with built-in SAM, then the RN-SAM corresponding to the RN-F will
be discovered as internal RN-SAM.

Add RN-F count as part of cmn600 context and use it in place of the
external RN-SAM count.

Change-Id: I9257e11b95b865eaf864ec8d9ba804c02e1ff2eb
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>