cmn600: fix hased memory region calculation in sam programming
In SAM programming sequence, region count for the SYS_CACHE_GRP_REGION
registers are not incremented. Having a second entry for
SYS_CACHE_GRP_REGION register in config data might overwrite the
existing entry or get programmed in a wrong region. To avoid this, use
`region_sys_count` as counter for the entries going to
SYS_CACHE_GRP_REGION register. This patch also adds bounds checking for
the region count of hashed and non-hased memory regions.

Change-Id: I32d0daf20790a91929c14285a42a03094a2eff6e
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>