DVFS: Add support for multiple concurrent SET_OPP requests
Enable multiple SET_OPP requests to be serviced as a single
'collapsed' request where the data from the last service request
overrides the earlier requests.

Some requests cannot be allowed to fail. This patch adds retry
functionality.

Always set frequency regardless of voltage. If the requested voltage
is equal to the actual voltage but the frequency has not yet been set
we go ahead and set the frequency.

Change-Id: I532bbd90104db399c275ecd55014c66e9b42d9ea
Signed-off-by: Jim Quigley <jim.quigley@arm.com>