i2c: Add support for concurrent requests
On a system, it is usual for several I2C slaves
to be attached to the same I2C bus. This results in
possible concurrent accesses to the I2C bus as a
firmware entity may want to access a device it
controls through the I2C bus while the bus is busy
communicating with another device.

This patch add support for concurrent requests on an
I2C bus. That way, users of this module don't have to
take into account that they may share an I2C bus.

This patch handles such concurrent accesses by
queueing I2C requests when received while the I2C bus
is busy.

The queueing and processing of the I2C request follow a
FIFO logic. When the processing of an I2C request is
completed, the processing of the request at the head of
the queue, if any, is initiated.

Change-Id: Ic9d6f1ac3bea9b16810f2a9252b24d4b33432025
Signed-off-by: Ronald Cron <ronald.cron@arm.com>