i2c: Introduce PANIC state
When processing an event for an I2C device, if an unrecoverable error
occurs, the I2C HAL internal state of the I2C device is set to the
PANIC state.

The possible type of unrecoverable errors are: unexpected event,
invalid event, call to framework returned in error. Driver errors
are not considered as unrecoverable.

It is not possible to exit the PANIC state. Once in PANIC state,
all requests returns with the FWK_E_PANIC error code.

Change-Id: I8d1a6b62cc1ab413e8a43e6641604133bb5dfbd8
Signed-off-by: Ronald Cron <ronald.cron@arm.com>