power_domain/shutdown: Add initial support for non-blocking calls
The call to pd_system_shutdown API in power_domain module is
reworked to replace put_event_and_wait to put_event. This change
allows callers running in the common thread to use that API.
Because of this change, pd_system_shutdown does return with either an
error or FWK_PENDING, the latter indicating the request has been
acknowledged.

For now, power_domain does not provide any response event when
returning the pending status, even if the shutdown would fail later.

As a consequence of this change, mod_scmi_system_power is accordingly
modified such that when the system shutdown request is acknowledged,
no response to the scmi agent is provided.

At last, this initial support for non-blocking call to system_system
can be seen as a fire-and-forget request.

Change-Id: If7289fb0985f7c4904f22cb81371836fe5cf5db3
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>