juno/clock: Bind HDLCD0 and HDLCD1 to CDCEL to expose response api
HDLCD driver can request CDCEL937 driver to perform a
set_rate_from_index. Since the response will be PENDING, it needs to
bind "as a HAL" to the CDCEL937 driver. So that it can receive the
request back from the driver, CDCEL937 needs to bind back to HDLCD for
HDLCD0 and HDLCD1 clock instead of binding back to the clock HAL.

Change-Id: I2753ee6c88f2ec8ec01f044e4e548d1aa3eabc19
Signed-off-by: Raphael Gault <raphael.gault@arm.com>