juno: Remove unused rate address from CDCEL937 clock driver
The CDCEL937 clock driver stores the address to return the clock rate in
the clock device context before it creates an event to read the rate.
If the get_rate() event is not created because another request is
already in progress the target address will have been overwritten.
Remove the unused rate address from asynchronous request code path.

Change-Id: I56da6ab26bf8f752a7d7500b6fd8fcf2fa45eca3
Signed-off-by: Jim Quigley <jim.quigley@arm.com>