juno: Fix clock-id handling in HDLCD driver
There is a bug in the HDLCD clock driver when it stores the ID of
the clock which is currently waiting for a PENDING request. The
clock ID is stored once for the module, so if requests are initiated
for multiple devices the clock ID of the last request will
overwrite the clock ID of the request which is actually in progress.
So we fix this by handling only one request at a time.

Change-Id: I9abbd572954b328da7123312f288a6cededeb98b
Signed-off-by: Jim Quigley <jim.quigley@arm.com>