power_domain: Fix pd in charge of response in process_set_state_request
This patch is partially reverting a previous change within
process_set_state_request and addressing correctly the missing
response when the driver call to set a new power state fails.

With this change, the power domain in charge of response
is assigned before any pre-state or state transitions happen.
This makes sure that in case of pre-state notifications
there is a power domain assigned for the response. In case
a state transition is initiated, if the call to the driver fails,
then the processing of the set state request is aborted and an
error returned immediately.

Change-Id: I7d084056bf10404f65287add3277db77f62d8618
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>