cmn600: use system info module to retrieve chip information
Commit 7acbd4d8fb4d ("cmn600: expose API to read chip information from
platforms") introduced an api to obtain chip information in the cmn600
module. As the system info module provides a reusable and scalable
interface to obtain system information, rework the cmn600 module to use
the system_info module to obtain the chip id and multi-chip information
and remove the chip info api in cmn600 and its implementation from the
n1sdp system module.

Change-Id: I2cfa45ba9bf4632f73bc021f4779f6bc2c96d93b
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>