power_domain: Fix exit from process_set_state_request
The function process_set_state_request can be called when
a response event is required or not, but currently the check
for this case is performed on the response event, and not the
event.

This patch slightly modifies process_set_state_request function
in order to correctly check the above condition.

Change-Id: I03023223925c35cf208949d0c0ef3fe7b8a34983
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>