cmn600: include encoding bits while calculation x and y position
Depending upon the mesh size (X and Y dimensions), CMN-600 supports
two different width for encoding the X and Y positions in the node id
value. If both X and Y size of the mesh is equal or less than 4, 2 bits
are used to encode the X and Y position in the node id. If X and Y size
are between 5 and 8 included, 3 bits are used to encode. This is not
considered while calculating the X and Y position from the node info
value which leads to wrong values if the encoding bits are of size 3.
This patch fixes the X and Y position calculation to include the
encoding bits.

Change-Id: I74ef3c39c4907f2a45aaf1023c981576f02c6539
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>