juno/cdcel937: Low precision should mean low precision when calculating preset.
If the preset is not found in the lookup table it is calculated using
the rate in MHz. To avoid floating point arithmetic and the associated
memory cost we will simplify the (rate / MHz) calculation to be a
uint32_t operation.

This results in a memory saving of approximately 1.5kb. As we should
usually get a hit in the lookup table this is an acceptable trade-off.

Change-Id: I2e2a78f85bdb3cd9e35149a324000935d3d8ebb5
Signed-off-by: Jim Quigley <jim.quigley@arm.com>