juno/xrp7724: Expand alarms to elements
In this module the alarm is used to trigger a deferred
event when processing a PSU set request and currently there is
only one alarm available to all elements. This is
a limitation since different devices can trigger the
same alarm, effectively restarting it, and overriding
any previously started. This reflects in PSU requests never
completed hence making the device unavailable.

This patch removes the alarm for the module and makes
available an alarm for each device.

Change-Id: Ie87601279ac7f48018fe9da3129b3f15a7235698
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>