cmn_rhodes: prepare module for multi-chip support
As preparation to add multi-chip support to CMN-Rhodes module, add
support for multiple device context table through elements. With this
addition, the configuration data of the CMN-Rhodes can have multiple
element table instead of single module data, each element index
indicating the device configuration data of a chip.

Note: This patch doesn't complete the multi-chip support, but merely
prepares the module for it.

Change-Id: I8cdf7d2708109c1ab0ddeb6ee6b03a53a23b8f79
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>