dbg: Add debugger CLI module
This module adds support for entering the CLI while running by
pressing CTRL-E.

It polls the UART in a non-blocking fashion using an alarm, and
fires an event if the correct key combination is detected, in order
to avoid running the CLI in the alarm ISR. Handling of further
events in the queue is blocked until the CLI is exited. Events
can still be queued, but will not be processed.

Implements support for N1SDP and SGM775 in release mode.

Change-Id: I4a20d71a9539935cba549e03ef4057ce6ddcb538
Signed-off-by: Anthony Roberts <anthony.roberts@arm.com>