cmn_rhodes: add chip-id and multi-chip mode to context
CMN-Rhodes interconnect can be built with Coherent Multichip Link (CML)
which enables support for Symmetric Multiprocessing (SMP) allowing
multiple chips to be connected in a coherent way. In such case,
CMN-Rhodes needs to configure the SAM table and setup the CML link based
on the chip-id information. Allow this module to obtain the multi-chip
mode and chip-id value from the system information module to facilitate
programming based on chip-id and multi-chip mode.

Change-Id: I710786229468a84ae0c6e377c735c201c201431a
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>