cmn_rhodes: update SAM programming based on chip-id
In a multi-chip scenario, the physical address space of each chip will
be offset based on the chip-id. For example, if address space of a
chip is 4TB, typically chip-0 physical address space will be between
0-4TB and chip-1 physical address space will be between 4TB-8TB.

Update RN-SAM and HN-F SAM programming based on the chip-id and the
address space.

Change-Id: I2f3d51d7e1045dce2d82c1909e3df9c773a16e18
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>