cmn_rhodes: fix cal mode register programming
The RNSAM_SYS_CACHE_GRP_CAL_MODE register can be configured to setup the
CAL mode for multiple System Cache Groups (SCG). But this register is
inadvertently overwritten when programming CAL mode for a SCG region
resulting in losing the CAL mode programming for other SCG regions. Fix
this by ensuring that previously programmed values are not lost.

Change-Id: Icff744f4881f2ab506d7219384098c57a1097e38
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>