cmn600: fix cal mode register programming
The RNSAM_SYS_CACHE_GRP_CAL_MODE register can be configured to setup the
CAL mode for multiple System Cache Groups (SCG). But this register is
inadvertently overwritten when programming CAL mode for a SCG region
resulting in losing the CAL mode programming for other SCG regions. Fix
this by ensuring that previously programmed values are not lost.

Change-Id: I8087d16dc111e53c8c0f23c69a5537d4bddb9f50
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>