rddaniel: change ram firmware module initialization order
In RD-Daniel, AP context area is in the shared internal trusted RAM
which is behind a SBSX node. Unless CMN-Rhodes is configured before
AP context module, accessing the context region will be forwarded to the
default HN-D node. To avoid this, reorder the modules to initialize the
CMN-Rhodes the before AP Context module. Clock and timer modules has to
be initialized before the CMN-Rhodes module, so move those module on top
of CMN-Rhodes module.

Change-Id: I9939ebace3c934331f6085d16bda5c1dddfc730c
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>