scmi_perf: Add support for performance statistics
Add support for performance statistics when they are present and
configured properly. This patch changes the default values sent over
SCMIv2 protocol to the AP regarding the statistics region. The new
addresses and length for the statistics shared memory are taken from
statistics module. The patch initializes the statistics module based on
performance domains information and uses statistics exposed API for
this.

Change-Id: I4f3da736c07828c6430c48e583f0168de81c7d3d
Signed-off-by: Lukasz Luba <lukasz.luba@arm.com>