fwk: Deprecate fwk_put_event_and_wait API for single-threaded firmware
This patch deprecates the fwk_put_event_and_wait API for
builds with BUILD_HAS_MULTITHREADING=no. As this is a
synchronous API it could have serious adverse effects on
performance and throughput when used in single-threaded
mode. Therefore, this API should not be used in the future
in single-threaded mode. All future development should
follow the deferred response architecture model.

Change-Id: I6eccdf4fd901509df14a57a7f6246790061f8409
Signed-off-by: Jim Quigley <jim.quigley@arm.com>