module: statistics: add basic functionality
The new module provides statistics which can be used by SCMI performance
or power domain protocols. The memory shared region exposes information
about usage of domains in the SCP, their current operation level and time
residency. The module manages shared memory using internal data
structures allowing to map only needed domains for a protocol. The
allocated shared memory is checked to not overflow during the
initialization based on platform configuration parameters. Some of the
data are duplicated and tracked in local structures to avoid potential
attacks from OSPM. The fields used for calculating offsets or addresses
in the SCP are not taken from the shared memory. The layout of the shared
memory statistic tables can be found in the SCMI v2 specification document
(DEN0056B).

Change-Id: I3e1efd01119d28aa1fa31ae961fb2ca3533b7255
Signed-off-by: Lukasz Luba <lukasz.luba@arm.com>